All .cc & .h merged
