+++ Makefile.am
@@ -36,8 +36,7 @@ dist_aclocal_DATA=
 TESTS+= ${check_PROGRAMS}
 
 include m4/include.am
-include src/ascore/include.am
-include src/asql/include.am
+include src/include.am
 include libattachsql-2.0/include.am
 include tests/include.am
 include examples/include.am
