--- src/asql/include.am
@@ -1,44 +1,0 @@
-# vim:ft=automake
-# included from Top Level Makefile.am
-# All paths should be given relative to the root
-
-noinst_HEADERS+= src/asql/common.h
-noinst_HEADERS+= src/asql/error_internal.h
-noinst_HEADERS+= src/asql/connect_internal.h
-noinst_HEADERS+= src/asql/query_internal.h
-noinst_HEADERS+= src/asql/statement_internal.h
-
-lib_LTLIBRARIES+= src/asql/libattachsql.la
-src_asql_libattachsql_la_SOURCES=
-src_asql_libattachsql_la_LIBADD=
-src_asql_libattachsql_la_LDFLAGS=
-src_asql_libattachsql_la_CFLAGS= -DBUILDING_ASQL
-src_asql_libattachsql_la_CXXFLAGS= -DBUILDING_ASQL
-
-if BUILD_WIN32
-src_asql_libattachsql_la_LIBADD+= -lmingw32
-src_asql_libattachsql_la_LIBADD+= -lws2_32
-src_asql_libattachsql_la_LIBADD+= -lpsapi
-src_asql_libattachsql_la_LIBADD+= -liphlpapi
-endif
-
-src_asql_libattachsql_la_SOURCES+= src/ascore/buffer.cc
-src_asql_libattachsql_la_SOURCES+= src/ascore/command.cc
-src_asql_libattachsql_la_SOURCES+= src/ascore/connect.cc
-src_asql_libattachsql_la_SOURCES+= src/ascore/sha1.cc
-src_asql_libattachsql_la_SOURCES+= src/ascore/net.cc
-src_asql_libattachsql_la_SOURCES+= src/ascore/pack.cc
-src_asql_libattachsql_la_SOURCES+= src/ascore/statement.cc
-
-src_asql_libattachsql_la_SOURCES+= src/asql/error.cc
-src_asql_libattachsql_la_SOURCES+= src/asql/connect.cc
-src_asql_libattachsql_la_SOURCES+= src/asql/pool.cc
-src_asql_libattachsql_la_SOURCES+= src/asql/query.cc
-src_asql_libattachsql_la_SOURCES+= src/asql/statement.cc
-src_asql_libattachsql_la_SOURCES+= src/asql/utility.cc
-
-src_asql_libattachsql_la_LDFLAGS+= -version-info ${LIBATTACHSQL_LIBRARY_VERSION}
-
-src_asql_libattachsql_la_LIBADD+= @LIBUV_LIB@
-src_asql_libattachsql_la_LIBADD+= @ZLIB_LIB@
-src_asql_libattachsql_la_LIBADD+= @OPENSSL_LIBS@
