+++ src/buffer.cc
