+++ src/buffer.h
