+++ src/connect.cc
@@ -20,19 +20,21 @@
 #include "connect.h"
 #include "sha1.h"
 #include "net.h"
+#include "query_internal.h"
 #include <errno.h>
 #include <string.h>
 #ifdef HAVE_OPENSSL
 # include <openssl/ssl.h>
 #endif
 
-attachsql_connect_t *attachsql_con_create(const char *host, in_port_t port, const char *user, const char *pass, const char *schema)
+attachsql_connect_t *attachsql_connect_create(const char *host, in_port_t port, const char *user, const char *pass, const char *schema, attachsql_error_t **error)
 {
   attachsql_connect_t *con;
 
   con = new (std::nothrow) attachsql_connect_t;
   if (con == NULL)
   {
+    attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_ALLOC, ATTACHSQL_ERROR_LEVEL_ERROR, "82100", "Allocation failure for connection object");
     return NULL;
   }
 
@@ -42,7 +44,9 @@ attachsql_connect_t *attachsql_con_create(const char *host, in_port_t port, cons
   {
     con->local_errcode= ASRET_USER_TOO_LONG;
     con->status= ATTACHSQL_CON_STATUS_PARAMETER_ERROR;
-    return con;
+    attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_PARAMETER, ATTACHSQL_ERROR_LEVEL_ERROR, "22023", "User name too long");
+    attachsql_connect_destroy(con);
+    return NULL;
   }
   con->user= user;
   // We don't really care how long pass is since we itterate though it during
@@ -52,18 +56,25 @@ attachsql_connect_t *attachsql_con_create(const char *host, in_port_t port, cons
   {
     con->local_errcode= ASRET_SCHEMA_TOO_LONG;
     con->status= ATTACHSQL_CON_STATUS_PARAMETER_ERROR;
-    return con;
+    attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_PARAMETER, ATTACHSQL_ERROR_LEVEL_ERROR, "22023", "Schema too long");
+    attachsql_connect_destroy(con);
+    return NULL;
   }
   con->schema= schema;
 
   return con;
 }
 
-void attachsql_con_destroy(attachsql_connect_t *con)
+void attachsql_connect_destroy(attachsql_connect_t *con)
 {
   if (con == NULL)
   {
     return;
+  }
+
+  if (con->in_query)
+  {
+    attachsql_query_close(con);
   }
 
   if (con->read_buffer != NULL)
@@ -155,7 +166,7 @@ void on_resolved(uv_getaddrinfo_t *resolver, int status, struct addrinfo *res)
   uv_freeaddrinfo(res);
 }
 
-attachsql_con_status_t attachsql_con_poll(attachsql_connect_t *con)
+attachsql_con_status_t attachsql_do_poll(attachsql_connect_t *con)
 {
   //asdebug("Connection poll");
   if (con == NULL)
@@ -176,6 +187,147 @@ attachsql_con_status_t attachsql_con_poll(attachsql_connect_t *con)
   attachsql_run_uv_loop(con);
 
   return con->status;
+}
+
+attachsql_return_t attachsql_connect_poll(attachsql_connect_t *con, attachsql_error_t **error)
+{
+  attachsql_con_status_t status;
+
+  if (con == NULL)
+  {
+    attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_PARAMETER, ATTACHSQL_ERROR_LEVEL_ERROR, "22023", "Invalid connection object");
+    return ATTACHSQL_RETURN_ERROR;
+  }
+  if (not con->in_pool)
+  {
+    status= attachsql_do_poll(con);
+  }
+  else
+  {
+    status= con->status;
+  }
+
+  switch (status)
+  {
+    case ATTACHSQL_CON_STATUS_PARAMETER_ERROR:
+      attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_PARAMETER, ATTACHSQL_ERROR_LEVEL_ERROR, "22023", "Bad parameter");
+      if (con->callback_fn != NULL)
+      {
+        con->callback_fn(con, ATTACHSQL_EVENT_ERROR, con->callback_context, *error);
+      }
+      return ATTACHSQL_RETURN_ERROR;
+      break;
+    case ATTACHSQL_CON_STATUS_NOT_CONNECTED:
+      return ATTACHSQL_RETURN_NOT_CONNECTED;
+      break;
+    case ATTACHSQL_CON_STATUS_CONNECTING:
+      return ATTACHSQL_RETURN_CONNECTING;
+      break;
+    case ATTACHSQL_CON_STATUS_CONNECT_FAILED:
+      if (con->local_errcode == ASRET_DNS_ERROR)
+      {
+        attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_HOST_UNKNOWN, ATTACHSQL_ERROR_LEVEL_ERROR, "08000", con->errmsg);
+      }
+      else if (con->server_errno != 0)
+      {
+        attachsql_error_server_create(con, error);
+      }
+      else
+      {
+        if (con->errmsg[0] != '\0')
+        {
+          attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_CONNECT, ATTACHSQL_ERROR_LEVEL_ERROR, "08000", con->errmsg);
+        }
+        else
+        {
+          attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_CONNECT, ATTACHSQL_ERROR_LEVEL_ERROR, "08000", "Unknown connection failure");
+        }
+      }
+      if (con->callback_fn != NULL)
+      {
+        con->callback_fn(con, ATTACHSQL_EVENT_ERROR, con->callback_context, *error);
+      }
+      return ATTACHSQL_RETURN_ERROR;
+      break;
+    case ATTACHSQL_CON_STATUS_BUSY:
+      return ATTACHSQL_RETURN_PROCESSING;
+      break;
+    case ATTACHSQL_CON_STATUS_SSL_ERROR:
+      attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_SSL, ATTACHSQL_ERROR_LEVEL_ERROR, "08000", con->errmsg);
+      if (con->callback_fn != NULL)
+      {
+        con->callback_fn(con, ATTACHSQL_EVENT_ERROR, con->callback_context, *error);
+      }
+      return ATTACHSQL_RETURN_ERROR;
+      break;
+    case ATTACHSQL_CON_STATUS_IDLE:
+      if (con->server_errno != 0)
+      {
+        // Server error during query
+        attachsql_error_server_create(con, error);
+        if (con->callback_fn != NULL)
+        {
+          con->callback_fn(con, ATTACHSQL_EVENT_ERROR, con->callback_context, *error);
+        }
+        return ATTACHSQL_RETURN_ERROR;
+      }
+      else if (con->command_status == ATTACHSQL_COMMAND_STATUS_EOF)
+      {
+        if (con->callback_fn != NULL)
+        {
+          con->callback_fn(con, ATTACHSQL_EVENT_EOF, con->callback_context, NULL);
+        }
+        con->all_rows_buffered= true;
+        con->command_status= ATTACHSQL_COMMAND_STATUS_EOF;
+        return ATTACHSQL_RETURN_EOF;
+      }
+      else if (con->command_status == ATTACHSQL_COMMAND_STATUS_ROW_IN_BUFFER)
+      {
+        if (con->buffer_rows)
+        {
+          return attachsql_query_row_buffer(con, error);
+        }
+        else
+        {
+          if (con->callback_fn != NULL)
+          {
+            con->callback_fn(con, ATTACHSQL_EVENT_ROW_READY, con->callback_context, NULL);
+          }
+          return ATTACHSQL_RETURN_ROW_READY;
+        }
+      }
+      else if (con->command_status == ATTACHSQL_COMMAND_STATUS_CONNECTED)
+      {
+        if (con->callback_fn != NULL)
+        {
+          con->callback_fn(con, ATTACHSQL_EVENT_CONNECTED, con->callback_context, NULL);
+        }
+        if (con->query_buffer_length > 0)
+        {
+          return attachsql_connect_query(con, error);
+        }
+        else
+        {
+          return ATTACHSQL_RETURN_EOF;
+        }
+      }
+      /* If we hit here something went wrong */
+      assert(0);
+      return ATTACHSQL_RETURN_EOF;
+      break;
+    case ATTACHSQL_CON_STATUS_NET_ERROR:
+      attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_SERVER_LOST, ATTACHSQL_ERROR_LEVEL_ERROR, "08006", con->errmsg);
+      if (con->callback_fn != NULL)
+      {
+        con->callback_fn(con, ATTACHSQL_EVENT_ERROR, con->callback_context, *error);
+      }
+      return ATTACHSQL_RETURN_ERROR;
+      break;
+  }
+  /* If we get this far 1. our compiler needs shooting for not catching a
+   * missing case statement and 2. something bad probably happened anyway */
+  attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_UNKNOWN, ATTACHSQL_ERROR_LEVEL_ERROR, "99999", "Unkown error occurred");
+  return ATTACHSQL_RETURN_ERROR;
 }
 
 void attachsql_check_for_data_cb(uv_check_t *handle, int status)
@@ -267,6 +419,164 @@ attachsql_con_status_t attachsql_do_connect(attachsql_connect_t *con)
   return con->status;
 }
 
+attachsql_return_t attachsql_connect_query(attachsql_connect_t *con, attachsql_error_t **error)
+{
+  attachsql_command_status_t ret;
+
+  if (con->query_buffer_statement)
+  {
+    attachsql_statement_prepare(con, con->query_buffer_length, con->query_buffer, NULL);
+    ret= con->command_status;
+  }
+  else
+  {
+    ret= attachsql_command_send(con, ATTACHSQL_COMMAND_QUERY, con->query_buffer, con->query_buffer_length);
+  }
+  if (ret == ATTACHSQL_COMMAND_STATUS_SEND_FAILED)
+  {
+    attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_SERVER_GONE, ATTACHSQL_ERROR_LEVEL_ERROR, "08006", con->errmsg);
+    return ATTACHSQL_RETURN_ERROR;
+  }
+  return ATTACHSQL_RETURN_PROCESSING;
+}
+
+bool attachsql_connect(attachsql_connect_t *con, attachsql_error_t **error)
+{
+  attachsql_con_status_t status;
+
+  /* TODO: Merge this in? */
+  status= attachsql_do_connect(con);
+
+  switch(status)
+  {
+    case ATTACHSQL_CON_STATUS_PARAMETER_ERROR:
+      attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_PARAMETER, ATTACHSQL_ERROR_LEVEL_ERROR, "22023", "Bad parameter");
+      if (con->callback_fn != NULL)
+      {
+        con->callback_fn(con, ATTACHSQL_EVENT_ERROR, con->callback_context, *error);
+      }
+      return false;
+      break;
+    case ATTACHSQL_CON_STATUS_NOT_CONNECTED:
+      /* Shouldn't happen */
+      return true;
+      break;
+    case ATTACHSQL_CON_STATUS_CONNECTING:
+      return true;
+      break;
+    case ATTACHSQL_CON_STATUS_CONNECT_FAILED:
+      if (con->local_errcode == ASRET_DNS_ERROR)
+      {
+        attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_HOST_UNKNOWN, ATTACHSQL_ERROR_LEVEL_ERROR, "08000", con->errmsg);
+      }
+      else if (con->server_errno != 0)
+      {
+        attachsql_error_server_create(con, error);
+      }
+      else
+      {
+        if (con->errmsg[0] != '\0')
+        {
+          attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_CONNECT, ATTACHSQL_ERROR_LEVEL_ERROR, "08000", con->errmsg);
+        }
+        else
+        {
+          attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_CONNECT, ATTACHSQL_ERROR_LEVEL_ERROR, "08000", "Unknown connection failure");
+        }
+      }
+      if (con->callback_fn != NULL)
+      {
+        con->callback_fn(con, ATTACHSQL_EVENT_ERROR, con->callback_context, *error);
+      }
+      return false;
+      break;
+    case ATTACHSQL_CON_STATUS_BUSY:
+      /* Should never be hit */
+      return false;
+      break;
+    case ATTACHSQL_CON_STATUS_IDLE:
+      if ((con->command_status == ATTACHSQL_COMMAND_STATUS_CONNECTED) and (con->callback_fn != NULL))
+      {
+        con->callback_fn(con, ATTACHSQL_EVENT_CONNECTED, con->callback_context, NULL);
+      }
+      return true;
+      break;
+    case ATTACHSQL_CON_STATUS_SSL_ERROR:
+      attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_SSL, ATTACHSQL_ERROR_LEVEL_ERROR, "08000", con->errmsg);
+      if (con->callback_fn != NULL)
+      {
+        con->callback_fn(con, ATTACHSQL_EVENT_ERROR, con->callback_context, *error);
+      }
+
+      return false;
+      break;
+    case ATTACHSQL_CON_STATUS_NET_ERROR:
+      attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_SERVER_LOST, ATTACHSQL_ERROR_LEVEL_ERROR, "08006", con->errmsg);
+      if (con->callback_fn != NULL)
+      {
+        con->callback_fn(con, ATTACHSQL_EVENT_ERROR, con->callback_context, *error);
+      }
+      return false;
+      break;
+  }
+
+  return true;
+}
+
+bool attachsql_connect_set_option(attachsql_connect_t *con, attachsql_options_t option, const void *arg)
+{
+  // arg option is for later
+  (void) arg;
+  if (con == NULL)
+  {
+    return false;
+  }
+
+  switch (option)
+  {
+    case ATTACHSQL_OPTION_COMPRESS:
+#ifdef HAVE_ZLIB
+      con->client_capabilities|= ATTACHSQL_CAPABILITY_COMPRESS;
+#else
+      return false;
+#endif
+      break;
+    case ATTACHSQL_OPTION_FOUND_ROWS:
+      con->client_capabilities|= ATTACHSQL_CAPABILITY_FOUND_ROWS;
+      break;
+    case ATTACHSQL_OPTION_IGNORE_SIGPIPE:
+      con->client_capabilities|= ATTACHSQL_CAPABILITY_IGNORE_SIGPIPE;
+      break;
+    case ATTACHSQL_OPTION_INTERACTIVE:
+      con->client_capabilities|= ATTACHSQL_CAPABILITY_INTERACTIVE;
+      break;
+    case ATTACHSQL_OPTION_LOCAL_FILES:
+      con->client_capabilities|= ATTACHSQL_CAPABILITY_LOCAL_FILES;
+      break;
+    case ATTACHSQL_OPTION_MULTI_STATEMENTS:
+      con->client_capabilities|= ATTACHSQL_CAPABILITY_MULTI_STATEMENTS;
+      break;
+    case ATTACHSQL_OPTION_NO_SCHEMA:
+      con->client_capabilities|= ATTACHSQL_CAPABILITY_NO_SCHEMA;
+      break;
+    case ATTACHSQL_OPTION_SSL_NO_VERIFY:
+#ifdef HAVE_OPENSSL
+      con->ssl.no_verify= true;
+#else
+      return false;
+#endif
+      break;
+    case ATTACHSQL_OPTION_SEMI_BLOCKING:
+      con->options.semi_block= true;
+      break;
+    case ATTACHSQL_OPTION_NONE:
+      return false;
+      break;
+    default:
+      return false;
+  }
+  return true;
+}
 
 void on_connect(uv_connect_t *req, int status)
 {
@@ -574,9 +884,40 @@ asret_t scramble_password(attachsql_connect_t *con, unsigned char *buffer)
   return ASRET_OK;
 }
 
-#ifdef HAVE_OPENSSL
-bool attachsql_con_set_ssl(attachsql_connect_t *con, const char *key, const char *cert, const char *ca, const char *capath, const char *cipher, bool verify)
+void attachsql_connect_set_callback(attachsql_connect_t *con, attachsql_callback_fn *function, void *context)
 {
+  con->callback_fn= function;
+  con->callback_context= context;
+}
+
+const char *attachsql_connect_get_server_version(attachsql_connect_t *con)
+{
+  if (con == NULL)
+  {
+    return NULL;
+  }
+
+  return con->server_version;
+}
+
+uint32_t attachsql_connect_get_connection_id(attachsql_connect_t *con)
+{
+  if (con == NULL)
+  {
+    return 0;
+  }
+  return con->thread_id;
+}
+
+#ifdef HAVE_OPENSSL
+bool attachsql_connect_set_ssl(attachsql_connect_t *con, const char *key, const char *cert, const char *ca, const char *capath, const char *cipher, bool verify, attachsql_error_t **error)
+{
+  if (con == NULL)
+  {
+    attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_PARAMETER, ATTACHSQL_ERROR_LEVEL_ERROR, "22023", "Connection parameter not valid");
+    return false;
+  }
+
   con->ssl.context= SSL_CTX_new(TLSv1_client_method());
 
   if (cipher != NULL)
@@ -585,6 +926,7 @@ bool attachsql_con_set_ssl(attachsql_connect_t *con, const char *key, const char
     {
       strncpy(con->errmsg, "Error setting SSL cipher list", ATTACHSQL_ERROR_BUFFER_SIZE - 1);
       con->errmsg[ATTACHSQL_ERROR_BUFFER_SIZE - 1]= '\0';
+      attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_SSL, ATTACHSQL_ERROR_LEVEL_ERROR, "22023", con->errmsg);
       return false;
     }
   }
@@ -593,6 +935,7 @@ bool attachsql_con_set_ssl(attachsql_connect_t *con, const char *key, const char
   {
     strncpy(con->errmsg, "Error loading the SSL certificate authority file", ATTACHSQL_ERROR_BUFFER_SIZE -1);
     con->errmsg[ATTACHSQL_ERROR_BUFFER_SIZE - 1]= '\0';
+    attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_SSL, ATTACHSQL_ERROR_LEVEL_ERROR, "22023", con->errmsg);
     return false;
   }
 
@@ -602,6 +945,7 @@ bool attachsql_con_set_ssl(attachsql_connect_t *con, const char *key, const char
     {
       strncpy(con->errmsg, "Error loading the SSL certificate file", ATTACHSQL_ERROR_BUFFER_SIZE - 1);
       con->errmsg[ATTACHSQL_ERROR_BUFFER_SIZE - 1]= '\0';
+      attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_SSL, ATTACHSQL_ERROR_LEVEL_ERROR, "22023", con->errmsg);
       return false;
     }
 
@@ -614,6 +958,7 @@ bool attachsql_con_set_ssl(attachsql_connect_t *con, const char *key, const char
     {
       strncpy(con->errmsg, "Cannot load the SSL key file", ATTACHSQL_ERROR_BUFFER_SIZE - 1);
       con->errmsg[ATTACHSQL_ERROR_BUFFER_SIZE - 1]= '\0';
+      attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_SSL, ATTACHSQL_ERROR_LEVEL_ERROR, "22023", con->errmsg);
       return false;
     }
 
@@ -621,6 +966,7 @@ bool attachsql_con_set_ssl(attachsql_connect_t *con, const char *key, const char
     {
       strncpy(con->errmsg, "Error validating the SSL private key", ATTACHSQL_ERROR_BUFFER_SIZE - 1);
       con->errmsg[ATTACHSQL_ERROR_BUFFER_SIZE - 1]= '\0';
+      attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_SSL, ATTACHSQL_ERROR_LEVEL_ERROR, "22023", con->errmsg);
       return false;
     }
   }
@@ -634,5 +980,19 @@ bool attachsql_con_set_ssl(attachsql_connect_t *con, const char *key, const char
   SSL_set_connect_state(con->ssl.ssl);
 
   return true;
+}
+#else
+bool attachsql_connect_set_ssl(attachsql_connect_t *con, const char *key, const char *cert, const char *ca, const char *capath, const char *cipher, bool verify, attachsql_error_t **error)
+{
+  (void) con;
+  (void) key;
+  (void) cert;
+  (void) ca;
+  (void) capath;
+  (void) cipher;
+  (void) verify;
+
+  attachsql_error_client_create(error, ATTACHSQL_ERROR_CODE_NO_SSL, ATTACHSQL_ERROR_LEVEL_ERROR, "22023", "SSL support has not been compiled in");
+  return false;
 }
 #endif
