+++ src/pack.h
