+++ src/sha1.cc
