+++ src/statement.h
@@ -2,7 +2,7 @@
  * Copyright 2014 Hewlett-Packard Development Company, L.P.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License. You may obtain 
+ * not use this file except in compliance with the License. You may obtain
  * a copy of the License at
  *
  *      http://www.apache.org/licenses/LICENSE-2.0
@@ -17,15 +17,21 @@
 
 #pragma once
 
-#include "debug.h"
-#include "return.h"
-#include "pack_macros.h"
-#include "pack.h"
-#include "constants.h"
-#include "structs.h"
-#include <new>
-#include <uv.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <assert.h>
+#include "config.h"
+#include "common.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+bool attachsql_stmt_execute(attachsql_stmt_st *stmt);
+
+bool attachsql_stmt_check_buffer_size(attachsql_stmt_st *stmt, size_t required);
+
+attachsql_command_status_t attachsql_stmt_fetch(attachsql_stmt_st *stmt);
+
+bool attachsql_statement_set_param(attachsql_connect_t *con, attachsql_column_type_t type, uint16_t param, size_t length, const void *value, bool is_unsigned, attachsql_error_t **error);
+
+#ifdef __cplusplus
+}
+#endif
